+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Removed obsolete support for `country_name`
+- Removed dependency on `league/iso3166`
 
 ## [3.0.2] - 2020-09-03
 ### Fixed